fix(middle_day && does_it_fit)
