+++ tests/middle_day_test/Cargo.toml
@@ -7,4 +7,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+chrono = "0.4.40"
 middle_day = { path = "../../solutions/middle_day"}
